+++ src/commands/command.js
@@ -30,7 +30,7 @@ export async function command(_subcommand, ...args) {
       'GETKEYSANDFLAGS <full-command>',
       '    Return the keys and the access flags from a full Redis command.',
       'HELP',
-      '    Prints this help.',
+      '    Print this help.',
     ]
   }
 
@@ -78,7 +78,9 @@ export async function command(_subcommand, ...args) {
     const { getKeyIndexes, exists } = await import('@ioredis/commands')
     const [cmd, ...subArgs] = args
     if (!exists(cmd)) {
-      throw new Error('ERR Invalid command specified')
+      throw new Error(
+        "ERR wrong number of arguments for 'command|getkeys' command"
+      )
     }
 
     return getKeyIndexes(cmd, subArgs).map(i => subArgs[i])