+++ src/commands/hscan.js
@@ -5,8 +5,9 @@ export function hscan(key, cursor, ...args) {
   if (!this.data.has(key)) {
     return ['0', []]
   }
-  const hKeys = Object.keys(this.data.get(key))
-  return scanHelper(hKeys, 1, cursor, ...args)
+  const entries = Object.entries(this.data.get(key))
+  const [cur, scannedEntries] = scanHelper(entries, 1, cursor, ...args)
+  return [cur, scannedEntries.flat()]
 }
 
 export function hscanBuffer(...args) {